Move `service/controller/clusterapi/v22/controllercontext`
to `service/controller/controllercontext`, replacing the existing
implementation there which was unused.
