+++ service/controller/clusterapi/v22/cluster_resource_set.go
@@ -22,7 +22,6 @@ import (
 	"sigs.k8s.io/cluster-api/pkg/client/clientset_generated/clientset"
 
 	"github.com/giantswarm/cluster-operator/pkg/project"
-	"github.com/giantswarm/cluster-operator/service/controller/clusterapi/v22/controllercontext"
 	"github.com/giantswarm/cluster-operator/service/controller/clusterapi/v22/key"
 	"github.com/giantswarm/cluster-operator/service/controller/clusterapi/v22/resource/app"
 	"github.com/giantswarm/cluster-operator/service/controller/clusterapi/v22/resource/certconfig"
@@ -37,6 +36,7 @@ import (
 	"github.com/giantswarm/cluster-operator/service/controller/clusterapi/v22/resource/tenantclients"
 	"github.com/giantswarm/cluster-operator/service/controller/clusterapi/v22/resource/updatemachinedeployments"
 	"github.com/giantswarm/cluster-operator/service/controller/clusterapi/v22/resource/workercount"
+	"github.com/giantswarm/cluster-operator/service/controller/controllercontext"
 )
 
 // ClusterResourceSetConfig contains necessary dependencies and settings for
