+++ service/controller/clusterapi/v22/resource/certconfig/desired.go
@@ -12,8 +12,8 @@ import (
 
 	"github.com/giantswarm/cluster-operator/pkg/label"
 	"github.com/giantswarm/cluster-operator/pkg/project"
-	"github.com/giantswarm/cluster-operator/service/controller/clusterapi/v22/controllercontext"
 	"github.com/giantswarm/cluster-operator/service/controller/clusterapi/v22/key"
+	"github.com/giantswarm/cluster-operator/service/controller/controllercontext"
 )
 
 // GetDesiredState returns all desired CertConfigs for managed certificates.
